Add memory allocators/deallocators.
These wrappers call standard malloc/free, or hbw_malloc()/hbw_free() if memkind
is available on the system.  The latter calls will attempt to allocate memory
in high-bandwidth memory, if such is available.

The resist_malloc() wrapper also aligns the memory on 64-byte cache line
boundaries. (This is the appropriate size for x86.) For other architectures the
cache line sizes may be different, and we will need to create a way to
distinguish among them, since one cannot obtain this value at run time.