Add Travis CI
This PR gives us the ability to use Travis for build tests.  I had to
add a from-source build for `uncrustify` to the `.travis.yml` file
because the version provided by Debian 12.04.  For the Clang build I
needed to add the `-lm` flag to the Makefile.  Work is needed to make
the gcc build proceed properly, but this PR is a start Travis CI for our
project.