resist-memory: fix allocators/deallocators for padded memory
In their original form in commit b7f888f ("Add memory
allocators/deallocators."), the deallocators did not work correctly and
could lead to segfaults. This commit fixes them by pointing to the
"base" memory (requested size + padding) with one pointer, and pointing
to the aligned location with a second pointer.