Modifications enabling ghost grid points
The velocity grid allocation is modified so that ghost grid points are
allocated.  The full grid including ghost zones is 3 times the size of
the real grid, based on considerations from the geometry of the line
forming region.  The real grid is of size `real_vr_count` and the full
grid including ghost zones is of size `vr_count`.  So, when iterating
over wavelengths in the `tau` and `src` arrays, the right chunk size is
`real_vr_count`.

Additionally only real grid cells need to have associated angles for
computing mean intensity.  So we don't make allocations for angles in
the ghost points.

Further specifics:

* Comment struct, add real velocity grid variable
* Comments explain choice of 3 for full grid factor
* Comments explain why angles aren't allocated for ghosts
* Changes to the output format to include real grid size