Enhance uncrustify experience
The `PASS` and `FAIL` messages emitted by `uncrustify` should now be
colored green and red respectively, making it easier to notice that a
formatting error has been committed.  This is a bit of hackery but
hopefully it will work portably (ha).  To make this all work I had to
set the `$SHELL` variable to `/bin/bash`.

An additional target `reformat` has been added that is a scorched-earth
target that reformats all the source files in place but preserves the
modification times.  If nothing is done it should be no change to the
source files.