Redefine and clarify wavelength parameters
This commit clarifies that there are two wavelength grids we have to
worry about, one for the eventual output spectrum and one for the
opacity bins.  They have different limits and in fact the former
determines the latter along with the LFR geometry.  To get the source
function kernel right we want to have the logic in that does that and we
ought to clarify the conventions earlier rather than later.  In more
detail what this commit does is:

* Replace `wl` limits in the config with spectrum limits and step
* Define opacity binning, limits from by spectrum limit and `max_vr`
* Update interface to clarify `spec_wl` and `opac_wl` grids
* Modify `config_init_default` to reflect this new interface
* Provide more sensible UVOIR default for `config_init_default`
* Update necessary assertions in `config`
* Define `opac_wl` params in `context` based on `spec_wl` lims, max_vr

This commit should be orthogonal to the one in #39, and once that one is
merged I'll rebase then merge this one.  I'll leave certain other
convention changes about `context` for a future commit.